fix: Restore camera layout support in `image` view
