+++ src/camera-manager/utils/camera-entity-from-config.ts
@@ -1,5 +1,7 @@
 import { CameraConfig } from '../../config/types';
 
-export const getCameraEntityFromConfig = (cameraConfig: CameraConfig): string | null => {
-  return cameraConfig.camera_entity ?? cameraConfig.webrtc_card?.entity ?? null;
+export const getCameraEntityFromConfig = (
+  cameraConfig?: CameraConfig,
+): string | null => {
+  return cameraConfig?.camera_entity ?? cameraConfig?.webrtc_card?.entity ?? null;
 };
